feat: DemoUI에서 비디오 선택 및 재생 상태 관리 로직을 DailyLecture로 이동 #23
